fix(web): Improve performance of simulator and architecture views
Having loads of components doesn't go well with Vue, and changing to
these views used to take seconds (in the case of the architecture view,
it even failed sometimes), so it had to be addressed...

For the simulator view, the problem was the registers. Having one
BPopover (with its BTableSimple) per register was just too much, so we
opted for using a generic modal, as with the memory view.

For the architecture view, using lazy tabs seemed to do the trick.