refactor!(executor): More generic interrupts
Interrupts now have two handlers: the CREATOR (old) handler, which only
handles syscalls (and can be defined in the architecture), and the
architecture/custom handler, which reads the information from the
architecture. By default, we use the CREATOR handler, but this can be
modified. Interrupts itself are also optional, for example MIPS doesn't
use them currently.

Other changes:
- Delete virtual PC
- Fix JSON schema URL containing trailing `#`